refactor(apps): Extract constants
