+++ src/ducks/apps/constants.js
@@ -1,0 +1,16 @@
+export const APP_TYPE = {
+  KONNECTOR: 'konnector',
+  WEBAPP: 'webapp'
+}
+
+export const REGISTRY_CHANNELS = {
+  DEV: 'dev',
+  BETA: 'beta',
+  STABLE: 'stable'
+}
+
+export const APP_STATE = {
+  READY: 'ready',
+  INSTALLING: 'installing',
+  ERRORED: 'errored'
+}
